Rename core package to gbd package
